26.10.23_2

+++ INTRA PTZ client/PelcoDE.cs
@@ -8,12 +8,12 @@ namespace INTRA_PTZ_client
 {
     public static class PelcoDE
     {
-        public static Byte[] getCommand(Byte address, Byte k1, Byte k2, Byte dt1, Byte dt2)
+        public static Byte[] getCommand(int address, Byte k1, Byte k2, Byte dt1, Byte dt2)
         {
             Byte[] command = new Byte[7];
 
             command[0] = 0xff;
-            command[1] = address;
+            command[1] = (Byte)address;
             command[2] = k1;
             command[3] = k2;
             command[4] = dt1;
